+++ tests/test_TimeseriesExtractor_additional.py
@@ -19,14 +19,5 @@ def test_TimeseriesExtractor_no_parallel_additional():
     pipeline_name = "fmriprep_1.0.0/fmriprep"
     extractor.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name, tr=1.2)
     
-    print(extractor.subject_timeseries, flush=True)
-
     assert extractor.subject_timeseries["01"]["run-0"].shape[-1] == 100
     assert extractor.subject_timeseries["01"]["run-0"].shape[0] == 40
-
-    old_subject_timeseries = extractor.subject_timeseries
-
-    extractor2 = TimeseriesExtractor(parcel_approach=parcel_approach, standardize="zscore_sample",
-                                    use_confounds=False)
-    
-    extractor2.get_bold(bids_dir=bids_dir, task="rest", pipeline_name=pipeline_name)